requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
flake8==3.9.1
future==0.18.2
idna==2.9
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
packaging==20.9
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
python-dotenv==0.13.0
python-telegram-bot==13.5
pytz==2021.1
requests==2.23.0
six==1.15.0
toml==0.10.2
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.9
